requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.9.1            # via flask-migrate
click==6.7
first==2.0.1              # via pip-tools
flask-marshmallow==0.7.0
flask-migrate==2.0.3
flask-script==2.0.5
flask-sqlalchemy==2.2
flask==0.12.1
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
kafka-python==1.3.3
mako==1.0.6               # via alembic
markupsafe==1.0           # via jinja2, mako
marshmallow-sqlalchemy==0.13.1
marshmallow==2.13.5
mysqlclient==1.3.10
numpy==1.12.1
pip-tools==1.9.0
py==1.4.33                # via pytest
pytest==3.0.7
python-editor==1.0.3      # via alembic
readline==6.2.4.1
six==1.10.0               # via flask-marshmallow, pip-tools
sqlalchemy==1.1.9
werkzeug==0.12.1          # via flask
